write in storage is necessary
